src/requirements.txt
antspyx==0.3.2
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
chart-studio==1.1.0
ci-info==0.2.0
click==8.1.3
connected-components-3d==3.10.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dipy==1.5.0
dominate==2.7.0
entrypoints==0.4
etelemetry==0.3.0
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.7.0
fonttools==4.33.3
h5py==3.6.0
idna==3.3
imageio==2.19.2
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isodate==0.6.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.8
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyter-server==1.17.0
jupyterlab==3.4.2
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
looseversion==1.0.1
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
monai==0.8.1
nbclassic==0.3.7
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.1
nibabel==3.2.2
nipype==1.8.1
notebook==6.4.11
notebook-shim==0.1.0
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
plotly==5.8.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
prov==2.0.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydot==1.4.2
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
rdflib==6.1.1
requests==2.27.1
retrying==1.3.3
scikit-image==0.19.2
scikit-learn==1.1.0
scipy==1.8.1
Send2Trash==1.8.0
SimpleITK==2.1.1.2
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
statsmodels==0.13.2
tenacity==8.0.1
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2022.5.4
tinycss2==1.1.1
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
traits==6.3.2
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websocket-client==1.3.2
widgetsnbextension==3.6.0
zipp==3.8.0
